fix: architecture-advisor-a2b33c4d (issues lane)
