+++ deploy/coterie-provision/src/install.rs
@@ -354,6 +354,127 @@ fn gather_inputs<P: Prompter>(
         || prompts.prompt_secret("First admin password (input hidden)"),
     )?;
 
+    let stripe = gather_stripe_inputs(args, prompts, no_prompt)?;
+
+    let discord = gather_discord_inputs(args, prompts, no_prompt)?;
+
+    let unifi = gather_unifi_inputs(args, prompts, no_prompt)?;
+
+    let enable_caddy = gather_caddy_inputs(args, prompts, no_prompt)?;
+
+    // Version selection. The version selector module fetches the
+    // release list separately in the production path; here we accept
+    // a flag/env override or fall back to a static prompt that just
+    // asks for a tag string. (The version-list UI lives in main.rs.)
+    let version = resolve(
+        "version",
+        "COTERIE_PROVISION_VERSION",
+        args.version.clone(),
+        None,
+        no_prompt,
+        || prompts.prompt_text("Coterie version tag to install (e.g. v1.1.0)", None),
+    )?;
+
+    // Idempotency check: if .env already exists, ask before overwriting.
+    let overwrite_env = if preflight.env_present {
+        let from_env = std::env::var("COTERIE_PROVISION_OVERWRITE_ENV")
+            .ok()
+            .map(|s| matches!(s.as_str(), "true" | "1" | "yes"))
+            .unwrap_or(false);
+        if from_env || args.overwrite_env {
+            true
+        } else if no_prompt {
+            return Err(anyhow!(
+                "/opt/coterie/.env already exists; pass --overwrite-env or set COTERIE_PROVISION_OVERWRITE_ENV=true to clobber"
+            ));
+        } else {
+            prompts.prompt_yn(
+                "/opt/coterie/.env exists. Overwrite with new values?",
+                false,
+            )?
+        }
+    } else {
+        true
+    };
+
+    // Generate a session secret (64 hex chars).
+    let session_secret = generate_session_secret();
+
+    Ok(ResolvedInputs {
+        org_name,
+        portal_domain,
+        marketing_domain,
+        contact_email,
+        admin_email,
+        admin_username,
+        admin_full_name,
+        admin_password,
+        enable_stripe: stripe.enable_stripe,
+        stripe_mode: stripe.stripe_mode,
+        stripe_publishable_key: stripe.stripe_publishable_key,
+        stripe_secret_key: stripe.stripe_secret_key,
+        stripe_webhook_secret: stripe.stripe_webhook_secret,
+        stripe_live_publishable_key: stripe.stripe_live_publishable_key,
+        stripe_live_secret_key: stripe.stripe_live_secret_key,
+        stripe_live_webhook_secret: stripe.stripe_live_webhook_secret,
+        enable_discord: discord.enable_discord,
+        discord_bot_token: discord.discord_bot_token,
+        discord_guild_id: discord.discord_guild_id,
+        discord_member_role_id: discord.discord_member_role_id,
+        discord_expired_role_id: discord.discord_expired_role_id,
+        enable_unifi: unifi.enable_unifi,
+        unifi_controller_url: unifi.unifi_controller_url,
+        unifi_username: unifi.unifi_username,
+        unifi_password: unifi.unifi_password,
+        unifi_site_id: unifi.unifi_site_id,
+        enable_caddy,
+        version,
+        overwrite_env,
+        session_secret,
+    })
+}
+
+/// Stripe configuration collected by the wizard, mirroring the Stripe
+/// fields of `ResolvedInputs`. Assembled into `ResolvedInputs` by
+/// `gather_inputs`.
+struct StripeInputs {
+    enable_stripe: bool,
+    stripe_mode: StripeMode,
+    stripe_publishable_key: Option<String>,
+    stripe_secret_key: Option<SecretString>,
+    stripe_webhook_secret: Option<SecretString>,
+    stripe_live_publishable_key: Option<String>,
+    stripe_live_secret_key: Option<SecretString>,
+    stripe_live_webhook_secret: Option<SecretString>,
+}
+
+/// Discord configuration collected by the wizard.
+struct DiscordInputs {
+    enable_discord: bool,
+    discord_bot_token: Option<SecretString>,
+    discord_guild_id: Option<String>,
+    discord_member_role_id: Option<String>,
+    discord_expired_role_id: Option<String>,
+}
+
+/// UniFi configuration collected by the wizard.
+struct UnifiInputs {
+    enable_unifi: bool,
+    unifi_controller_url: Option<String>,
+    unifi_username: Option<String>,
+    unifi_password: Option<SecretString>,
+    unifi_site_id: Option<String>,
+}
+
+/// Collect Stripe inputs: the enable gate, then (when enabled) the
+/// test/live mode branch and the optional live-credential pre-load.
+/// Flag -> env -> prompt resolution order and every prefix check are
+/// unchanged from the inline version.
+fn gather_stripe_inputs<P: Prompter>(
+    args: &InstallArgs,
+    prompts: &P,
+    no_prompt: bool,
+) -> Result<StripeInputs> {
     let enable_stripe = resolve_bool(
         "enable-stripe",
         "COTERIE_PROVISION_ENABLE_STRIPE",
@@ -526,6 +647,25 @@ fn gather_inputs<P: Prompter>(
         (StripeMode::Live, None, None, None, None, None, None)
     };
 
+    Ok(StripeInputs {
+        enable_stripe,
+        stripe_mode,
+        stripe_publishable_key: stripe_pk,
+        stripe_secret_key: stripe_sk,
+        stripe_webhook_secret: stripe_whsec,
+        stripe_live_publishable_key: stripe_live_pk,
+        stripe_live_secret_key: stripe_live_sk,
+        stripe_live_webhook_secret: stripe_live_whsec,
+    })
+}
+
+/// Collect Discord inputs: the enable gate, then (when enabled) the bot
+/// token, guild ID, and role IDs.
+fn gather_discord_inputs<P: Prompter>(
+    args: &InstallArgs,
+    prompts: &P,
+    no_prompt: bool,
+) -> Result<DiscordInputs> {
     let enable_discord = resolve_bool(
         "enable-discord",
         "COTERIE_PROVISION_ENABLE_DISCORD",
@@ -574,6 +714,22 @@ fn gather_inputs<P: Prompter>(
             (None, None, None, None)
         };
 
+    Ok(DiscordInputs {
+        enable_discord,
+        discord_bot_token: discord_token,
+        discord_guild_id: discord_guild,
+        discord_member_role_id: discord_member_role,
+        discord_expired_role_id: discord_expired_role,
+    })
+}
+
+/// Collect UniFi inputs: the enable gate, then (when enabled) the
+/// controller URL, credentials, and site ID.
+fn gather_unifi_inputs<P: Prompter>(
+    args: &InstallArgs,
+    prompts: &P,
+    no_prompt: bool,
+) -> Result<UnifiInputs> {
     let enable_unifi = resolve_bool(
         "enable-unifi",
         "COTERIE_PROVISION_ENABLE_UNIFI",
@@ -620,85 +776,29 @@ fn gather_inputs<P: Prompter>(
         (None, None, None, None)
     };
 
-    let enable_caddy = resolve_bool(
+    Ok(UnifiInputs {
+        enable_unifi,
+        unifi_controller_url: unifi_url,
+        unifi_username: unifi_user,
+        unifi_password: unifi_pw,
+        unifi_site_id: unifi_site,
+    })
+}
+
+/// Collect the Caddy enable gate.
+fn gather_caddy_inputs<P: Prompter>(
+    args: &InstallArgs,
+    prompts: &P,
+    no_prompt: bool,
+) -> Result<bool> {
+    resolve_bool(
         "enable-caddy",
         "COTERIE_PROVISION_ENABLE_CADDY",
         args.enable_caddy,
         Some(true),
         no_prompt,
         || prompts.prompt_yn("Install and configure Caddy?", true),
-    )?;
-
-    // Version selection. The version selector module fetches the
-    // release list separately in the production path; here we accept
-    // a flag/env override or fall back to a static prompt that just
-    // asks for a tag string. (The version-list UI lives in main.rs.)
-    let version = resolve(
-        "version",
-        "COTERIE_PROVISION_VERSION",
-        args.version.clone(),
-        None,
-        no_prompt,
-        || prompts.prompt_text("Coterie version tag to install (e.g. v1.1.0)", None),
-    )?;
-
-    // Idempotency check: if .env already exists, ask before overwriting.
-    let overwrite_env = if preflight.env_present {
-        let from_env = std::env::var("COTERIE_PROVISION_OVERWRITE_ENV")
-            .ok()
-            .map(|s| matches!(s.as_str(), "true" | "1" | "yes"))
-            .unwrap_or(false);
-        if from_env || args.overwrite_env {
-            true
-        } else if no_prompt {
-            return Err(anyhow!(
-                "/opt/coterie/.env already exists; pass --overwrite-env or set COTERIE_PROVISION_OVERWRITE_ENV=true to clobber"
-            ));
-        } else {
-            prompts.prompt_yn(
-                "/opt/coterie/.env exists. Overwrite with new values?",
-                false,
-            )?
-        }
-    } else {
-        true
-    };
-
-    // Generate a session secret (64 hex chars).
-    let session_secret = generate_session_secret();
-
-    Ok(ResolvedInputs {
-        org_name,
-        portal_domain,
-        marketing_domain,
-        contact_email,
-        admin_email,
-        admin_username,
-        admin_full_name,
-        admin_password,
-        enable_stripe,
-        stripe_mode,
-        stripe_publishable_key: stripe_pk,
-        stripe_secret_key: stripe_sk,
-        stripe_webhook_secret: stripe_whsec,
-        stripe_live_publishable_key: stripe_live_pk,
-        stripe_live_secret_key: stripe_live_sk,
-        stripe_live_webhook_secret: stripe_live_whsec,
-        enable_discord,
-        discord_bot_token: discord_token,
-        discord_guild_id: discord_guild,
-        discord_member_role_id: discord_member_role,
-        discord_expired_role_id: discord_expired_role,
-        enable_unifi,
-        unifi_controller_url: unifi_url,
-        unifi_username: unifi_user,
-        unifi_password: unifi_pw,
-        unifi_site_id: unifi_site,
-        enable_caddy,
-        version,
-        overwrite_env,
-        session_secret,
-    })
+    )
 }
 
 fn generate_session_secret() -> SecretString {
