+++ openspec/issues/archive/2026-06-17-architecture-advisor-a2b33c4d/issue.md
