+++ src/service/event_admin_service.rs
@@ -649,1409 +649,5 @@ impl EventAdminService {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::{
-        domain::{CreateMemberRequest, EventType, EventVisibility, Recurrence, WeekdayCode},
-        integrations::IntegrationManager,
-        repository::{
-            MemberRepository, SqliteEventRepository, SqliteEventSeriesRepository,
-            SqliteMemberRepository,
-        },
-    };
-    use chrono::{Datelike, Duration, Weekday};
-    use sqlx::{Executor, SqlitePool};
-
-    /// Next Saturday at 18:00 UTC strictly after `now + 1 day`. Used as
-    /// the start time for single-event tests that don't care about the
-    /// weekday — just need a valid future timestamp.
-    fn next_saturday_anchor() -> DateTime<Utc> {
-        let now = Utc::now();
-        let start = now + Duration::days(1);
-        let days_until_sat = (Weekday::Sat.num_days_from_monday() as i64
-            - start.weekday().num_days_from_monday() as i64)
-            .rem_euclid(7);
-        let date = start.date_naive() + Duration::days(days_until_sat);
-        date.and_hms_opt(18, 0, 0).unwrap().and_utc()
-    }
-
-    /// Next Tuesday at 18:00 UTC strictly after `now + 1 day`. Used as
-    /// the start time for recurring-Tuesday tests where the weekly rule
-    /// requires the anchor BE a Tuesday.
-    fn next_tuesday_anchor() -> DateTime<Utc> {
-        let now = Utc::now();
-        let start = now + Duration::days(1);
-        let days_until_tue = (Weekday::Tue.num_days_from_monday() as i64
-            - start.weekday().num_days_from_monday() as i64)
-            .rem_euclid(7);
-        let date = start.date_naive() + Duration::days(days_until_tue);
-        date.and_hms_opt(18, 0, 0).unwrap().and_utc()
-    }
-
-    /// `anchor` shifted forward by `weeks` whole weeks. Use to compute
-    /// `until_date` values relative to the test's local anchor.
-    fn weeks_after(anchor: DateTime<Utc>, weeks: i64) -> DateTime<Utc> {
-        anchor + Duration::weeks(weeks)
-    }
-
-    async fn fresh_pool() -> SqlitePool {
-        let pool = sqlx::sqlite::SqlitePoolOptions::new()
-            .max_connections(1)
-            .after_connect(|conn, _| {
-                Box::pin(async move {
-                    conn.execute("PRAGMA foreign_keys = ON").await?;
-                    Ok(())
-                })
-            })
-            .connect("sqlite::memory:")
-            .await
-            .expect(":memory:");
-        sqlx::migrate!("./migrations")
-            .run(&pool)
-            .await
-            .expect("migrate");
-        pool
-    }
-
-    fn make_service(pool: SqlitePool) -> EventAdminService {
-        let event_repo: Arc<dyn EventRepository> =
-            Arc::new(SqliteEventRepository::new(pool.clone()));
-        let series_repo: Arc<dyn EventSeriesRepository> =
-            Arc::new(SqliteEventSeriesRepository::new(pool.clone()));
-        let recurring = Arc::new(RecurringEventService::new(
-            event_repo.clone(),
-            series_repo.clone(),
-            pool.clone(),
-        ));
-        let audit = Arc::new(AuditService::new(pool.clone()));
-        let integrations = Arc::new(IntegrationManager::new());
-
-        EventAdminService::new(event_repo, series_repo, recurring, audit, integrations)
-    }
-
-    async fn make_actor(pool: &SqlitePool) -> Uuid {
-        let repo = SqliteMemberRepository::new(pool.clone());
-        let m = repo
-            .create(CreateMemberRequest {
-                email: format!("a-{}@example.com", Uuid::new_v4()),
-                username: format!("u_{}", Uuid::new_v4().simple()),
-                full_name: "Test Admin".to_string(),
-                password: "p4ssword_long_enough".to_string(),
-                membership_type_id: None,
-                ..Default::default()
-            })
-            .await
-            .unwrap();
-        m.id
-    }
-
-    async fn audit_count(pool: &SqlitePool, action: &str, entity_id: &str) -> i64 {
-        let count: (i64,) =
-            sqlx::query_as("SELECT COUNT(*) FROM audit_logs WHERE action = ? AND entity_id = ?")
-                .bind(action)
-                .bind(entity_id)
-                .fetch_one(pool)
-                .await
-                .unwrap();
-        count.0
-    }
-
-    fn single_input(start: DateTime<Utc>, visibility: EventVisibility) -> CreateEventInput {
-        CreateEventInput {
-            title: "Test Event".to_string(),
-            description: "A test event".to_string(),
-            event_type: EventType::Meeting,
-            event_type_id: None,
-            visibility,
-            start_time: start,
-            end_time: None,
-            location: None,
-            max_attendees: None,
-            rsvp_required: false,
-            image_url: None,
-            recurrence: None,
-            recurrence_until: None,
-        }
-    }
-
-    #[tokio::test]
-    async fn create_single_event_emits_full_chain() {
-        let pool = fresh_pool().await;
-        let svc = make_service(pool.clone());
-        let actor = make_actor(&pool).await;
-
-        let start = next_saturday_anchor();
-        let input = single_input(start, EventVisibility::MembersOnly);
-
-        let event = svc.create(actor, input).await.unwrap();
-
-        // Repo touched — event exists and matches input.
-        let fetched = svc.event_repo.find_by_id(event.id).await.unwrap().unwrap();
-        assert_eq!(fetched.title, "Test Event");
-        assert_eq!(fetched.visibility, EventVisibility::MembersOnly);
-        assert!(
-            fetched.series_id.is_none(),
-            "non-recurring create should not set series_id"
-        );
-
-        // Audit row inserted.
-        assert_eq!(
-            audit_count(&pool, "create_event", &event.id.to_string()).await,
-            1
-        );
-
-        // No series row created (single insert).
-        let series_count: (i64,) = sqlx::query_as("SELECT COUNT(*) FROM event_series")
-            .fetch_one(&pool)
-            .await
-            .unwrap();
-        assert_eq!(series_count.0, 0);
-
-        // Integration dispatch fired — IntegrationManager has no
-        // registered integrations in this test so the call is a no-op
-        // but reaching here without panic confirms the chain ran.
-    }
-
-    #[tokio::test]
-    async fn create_recurring_series_materializes_and_audits() {
-        let pool = fresh_pool().await;
-        let svc = make_service(pool.clone());
-        let actor = make_actor(&pool).await;
-
-        let start = next_tuesday_anchor();
-        let until = weeks_after(start, 8);
-        let mut input = single_input(start, EventVisibility::MembersOnly);
-        input.recurrence = Some(Recurrence::WeeklyByDay {
-            interval: 1,
-            weekdays: vec![WeekdayCode::Tue],
-        });
-        input.recurrence_until = Some(until);
-
-        let anchor = svc.create(actor, input).await.unwrap();
-
-        // Anchor is an actual occurrence row.
-        assert!(anchor.series_id.is_some());
-        let series_id = anchor.series_id.unwrap();
-
-        // Multiple occurrences materialized.
-        let count: (i64,) = sqlx::query_as("SELECT COUNT(*) FROM events WHERE series_id = ?")
-            .bind(series_id.to_string())
-            .fetch_one(&pool)
-            .await
-            .unwrap();
-        assert!(
-            count.0 > 1,
-            "expected multiple occurrences, got {}",
-            count.0
-        );
-
-        // Series row exists.
-        let series_count: (i64,) = sqlx::query_as("SELECT COUNT(*) FROM event_series WHERE id = ?")
-            .bind(series_id.to_string())
-            .fetch_one(&pool)
-            .await
-            .unwrap();
-        assert_eq!(series_count.0, 1);
-
-        // Audit row uses create_event_series with series_id as the entity_id.
-        assert_eq!(
-            audit_count(&pool, "create_event_series", &series_id.to_string()).await,
-            1,
-        );
-        // And NOT a per-occurrence create_event audit row.
-        assert_eq!(
-            audit_count(&pool, "create_event", &anchor.id.to_string()).await,
-            0
-        );
-    }
-
-    #[tokio::test]
-    async fn create_admin_only_event_skips_integration_dispatch() {
-        // No external observability of the integration_manager call
-        // beyond "did the method return Ok"; this test ensures the
-        // AdminOnly branch persists the event + audit row without
-        // panicking and asserts the visibility was preserved end-to-end.
-        let pool = fresh_pool().await;
-        let svc = make_service(pool.clone());
-        let actor = make_actor(&pool).await;
-
-        let start = next_saturday_anchor();
-        let input = single_input(start, EventVisibility::AdminOnly);
-
-        let event = svc.create(actor, input).await.unwrap();
-        assert_eq!(event.visibility, EventVisibility::AdminOnly);
-        assert_eq!(
-            audit_count(&pool, "create_event", &event.id.to_string()).await,
-            1
-        );
-    }
-
-    fn update_input_from(event: &Event) -> UpdateEventInput {
-        UpdateEventInput {
-            title: event.title.clone(),
-            description: event.description.clone(),
-            event_type: event.event_type.clone(),
-            event_type_id: event.event_type_id,
-            visibility: event.visibility.clone(),
-            start_time: event.start_time,
-            end_time: event.end_time,
-            location: event.location.clone(),
-            max_attendees: event.max_attendees,
-            rsvp_required: event.rsvp_required,
-            image_url: event.image_url.clone(),
-        }
-    }
-
-    #[tokio::test]
-    async fn update_one_writes_audit() {
-        let pool = fresh_pool().await;
-        let svc = make_service(pool.clone());
-        let actor = make_actor(&pool).await;
-
-        let start = next_saturday_anchor();
-        let event = svc
-            .create(actor, single_input(start, EventVisibility::MembersOnly))
-            .await
-            .unwrap();
-
-        let mut input = update_input_from(&event);
-        input.title = "Renamed".to_string();
-        let result = svc.update_one(actor, event.id, input).await.unwrap();
-
-        assert_eq!(result.title, "Renamed");
-        assert_eq!(
-            audit_count(&pool, "update_event", &event.id.to_string()).await,
-            1
-        );
-    }
-
-    #[tokio::test]
-    async fn update_series_from_audits_with_count() {
-        let pool = fresh_pool().await;
-        let svc = make_service(pool.clone());
-        let actor = make_actor(&pool).await;
-
-        let start = next_tuesday_anchor();
-        let until = weeks_after(start, 8);
-        let mut input = single_input(start, EventVisibility::MembersOnly);
-        input.recurrence = Some(Recurrence::WeeklyByDay {
-            interval: 1,
-            weekdays: vec![WeekdayCode::Tue],
-        });
-        input.recurrence_until = Some(until);
-
-        let anchor = svc.create(actor, input).await.unwrap();
-        let series_id = anchor.series_id.unwrap();
-
-        let mut update = update_input_from(&anchor);
-        update.title = "Renamed Series".to_string();
-        let count = svc
-            .update_series_from(actor, series_id, anchor.start_time, update)
-            .await
-            .unwrap();
-        assert!(count >= 1);
-
-        assert_eq!(
-            audit_count(&pool, "update_event_series", &series_id.to_string()).await,
-            1,
-        );
-
-        // Anchor row reflects the new title.
-        let fetched = svc.event_repo.find_by_id(anchor.id).await.unwrap().unwrap();
-        assert_eq!(fetched.title, "Renamed Series");
-    }
-
-    #[tokio::test]
-    async fn delete_one_writes_audit() {
-        let pool = fresh_pool().await;
-        let svc = make_service(pool.clone());
-        let actor = make_actor(&pool).await;
-
-        let start = next_saturday_anchor();
-        let event = svc
-            .create(actor, single_input(start, EventVisibility::MembersOnly))
-            .await
-            .unwrap();
-
-        svc.delete_one(actor, event.id).await.unwrap();
-        assert!(svc.event_repo.find_by_id(event.id).await.unwrap().is_none());
-        assert_eq!(
-            audit_count(&pool, "delete_event", &event.id.to_string()).await,
-            1
-        );
-    }
-
-    #[tokio::test]
-    async fn end_series_caps_and_audits() {
-        let pool = fresh_pool().await;
-        let svc = make_service(pool.clone());
-        let actor = make_actor(&pool).await;
-
-        let start = next_tuesday_anchor();
-        let until = weeks_after(start, 8);
-        let mut input = single_input(start, EventVisibility::MembersOnly);
-        input.recurrence = Some(Recurrence::WeeklyByDay {
-            interval: 1,
-            weekdays: vec![WeekdayCode::Tue],
-        });
-        input.recurrence_until = Some(until);
-
-        let anchor = svc.create(actor, input).await.unwrap();
-        let series_id = anchor.series_id.unwrap();
-
-        let before: (i64,) = sqlx::query_as("SELECT COUNT(*) FROM events WHERE series_id = ?")
-            .bind(series_id.to_string())
-            .fetch_one(&pool)
-            .await
-            .unwrap();
-
-        // End the series at anchor.start_time — should delete all
-        // strictly-later occurrences and leave anchor.
-        let removed = svc
-            .end_series(actor, series_id, anchor.start_time)
-            .await
-            .unwrap();
-        assert!(removed > 0);
-
-        let after: (i64,) = sqlx::query_as("SELECT COUNT(*) FROM events WHERE series_id = ?")
-            .bind(series_id.to_string())
-            .fetch_one(&pool)
-            .await
-            .unwrap();
-        assert_eq!(after.0, before.0 - removed as i64);
-
-        // Anchor still there.
-        assert!(svc
-            .event_repo
-            .find_by_id(anchor.id)
-            .await
-            .unwrap()
-            .is_some());
-
-        // until_date set on the series row.
-        let series = svc
-            .event_series_repo
-            .find_by_id(series_id)
-            .await
-            .unwrap()
-            .unwrap();
-        assert_eq!(series.until_date, Some(anchor.start_time));
-
-        assert_eq!(
-            audit_count(&pool, "end_series", &series_id.to_string()).await,
-            1,
-        );
-    }
-
-    // -----------------------------------------------------------------
-    // Per-occurrence exception tests
-
-    use crate::domain::{OccurrenceException, OccurrenceExceptionKind, OccurrenceOverride};
-
-    fn recurring_input(start: DateTime<Utc>, until: DateTime<Utc>) -> CreateEventInput {
-        let mut input = single_input(start, EventVisibility::MembersOnly);
-        input.recurrence = Some(Recurrence::WeeklyByDay {
-            interval: 1,
-            weekdays: vec![WeekdayCode::Tue],
-        });
-        input.recurrence_until = Some(until);
-        input
-    }
-
-    async fn make_series_with_anchor(svc: &EventAdminService, actor: Uuid) -> (Uuid, Event) {
-        let start = next_tuesday_anchor();
-        let until = weeks_after(start, 12);
-        let anchor = svc
-            .create(actor, recurring_input(start, until))
-            .await
-            .unwrap();
-        (anchor.series_id.unwrap(), anchor)
-    }
-
-    #[tokio::test]
-    async fn cancel_event_occurrence_writes_exception_and_deletes_row() {
-        let pool = fresh_pool().await;
-        let svc = make_service(pool.clone());
-        let actor = make_actor(&pool).await;
-        let (series_id, _anchor) = make_series_with_anchor(&svc, actor).await;
-
-        // Cancel occurrence 3 (a future occurrence).
-        let before = svc
-            .event_repo
-            .find_by_series_and_index(series_id, 3)
-            .await
-            .unwrap();
-        assert!(before.is_some(), "occurrence 3 should exist pre-cancel");
-
-        svc.cancel_event_occurrence(actor, series_id, 3, Some("holiday".into()))
-            .await
-            .unwrap();
-
-        // Events row is gone.
-        let after = svc
-            .event_repo
-            .find_by_series_and_index(series_id, 3)
-            .await
-            .unwrap();
-        assert!(
-            after.is_none(),
-            "occurrence 3 should be deleted after cancel"
-        );
-
-        // Exception row exists.
-        let ex = svc
-            .event_series_repo
-            .find_exception(series_id, 3)
-            .await
-            .unwrap();
-        let ex = ex.expect("exception should be present");
-        assert_eq!(ex.kind, OccurrenceExceptionKind::Cancelled);
-        assert_eq!(ex.audit_reason.as_deref(), Some("holiday"));
-
-        // Audit entry.
-        assert!(
-            audit_count(
-                &pool,
-                "cancel_event_occurrence",
-                &before.unwrap().id.to_string()
-            )
-            .await
-                >= 1,
-            "expected cancel_event_occurrence audit",
-        );
-    }
-
-    #[tokio::test]
-    async fn cancelled_occurrence_does_not_reappear_after_materializer_run() {
-        let pool = fresh_pool().await;
-        let svc = make_service(pool.clone());
-        let actor = make_actor(&pool).await;
-        let (series_id, _anchor) = make_series_with_anchor(&svc, actor).await;
-
-        svc.cancel_event_occurrence(actor, series_id, 4, None)
-            .await
-            .unwrap();
-
-        // Run the materializer over the same series. The horizon-roll
-        // is a no-op when materialized_through is already at the target,
-        // so force a re-materialization by calling extend_horizon
-        // directly with a target beyond the current horizon.
-        let series = svc
-            .event_series_repo
-            .find_by_id(series_id)
-            .await
-            .unwrap()
-            .unwrap();
-        let new_target = series.materialized_through + chrono::Duration::weeks(20);
-        svc.recurring_event_service
-            .extend_horizon(&series, new_target)
-            .await
-            .unwrap();
-
-        // Occurrence 4 should still be absent.
-        let still_gone = svc
-            .event_repo
-            .find_by_series_and_index(series_id, 4)
-            .await
-            .unwrap();
-        assert!(still_gone.is_none(), "cancelled occurrence reappeared");
-
-        // Exception row still present.
-        let ex = svc
-            .event_series_repo
-            .find_exception(series_id, 4)
-            .await
-            .unwrap();
-        assert!(ex.is_some());
-    }
-
-    /// Regression for stream-index misalignment after a boundary cancel.
-    ///
-    /// Cancelling the highest-numbered materialized occurrence hard-deletes
-    /// its `events` row but leaves `materialized_through` unchanged. The
-    /// horizon-roll must keep numbering by stream position (not by
-    /// `MAX(occurrence_index)` of survivors); otherwise the first new
-    /// occurrence is mis-assigned the cancelled index, collides with that
-    /// `Cancelled` exception, and is silently skipped — shifting every
-    /// later index down by one.
-    #[tokio::test]
-    async fn extend_horizon_after_cancelling_boundary_occurrence_keeps_indices_aligned() {
-        let pool = fresh_pool().await;
-        let svc = make_service(pool.clone());
-        let actor = make_actor(&pool).await;
-
-        // Materialize EXACTLY occurrences 1..=5; occurrence 5 is the
-        // horizon boundary (N = 5). For weekly-by-Tuesday, occurrence k
-        // starts at anchor + (k-1) weeks, so an until_date one day past
-        // occurrence 5 includes it and stops before occurrence 6.
-        let start = next_tuesday_anchor();
-        let until = start + Duration::weeks(4) + Duration::days(1);
-        let anchor = svc
-            .create(actor, recurring_input(start, until))
-            .await
-            .unwrap();
-        let series_id = anchor.series_id.unwrap();
-        let initial: (i64,) = sqlx::query_as("SELECT COUNT(*) FROM events WHERE series_id = ?")
-            .bind(series_id.to_string())
-            .fetch_one(&pool)
-            .await
-            .unwrap();
-        assert_eq!(
-            initial.0, 5,
-            "expected exactly 5 occurrences (1..=5) materialized"
-        );
-
-        // Open the series back up so extend_horizon can roll past the cap.
-        sqlx::query("UPDATE event_series SET until_date = NULL WHERE id = ?")
-            .bind(series_id.to_string())
-            .execute(&pool)
-            .await
-            .unwrap();
-
-        // Cancel the boundary occurrence (N = 5): deletes its events row,
-        // writes a Cancelled exception at index 5, leaves
-        // materialized_through unchanged. MAX(occurrence_index) now drops
-        // to 4 while the true stream boundary is still 5.
-        svc.cancel_event_occurrence(actor, series_id, 5, Some("boundary".into()))
-            .await
-            .unwrap();
-        assert!(
-            svc.event_repo
-                .find_by_series_and_index(series_id, 5)
-                .await
-                .unwrap()
-                .is_none(),
-            "boundary occurrence 5 should be deleted after cancel"
-        );
-
-        // Roll the horizon forward to cover stream positions 6..=10.
-        let series = svc
-            .event_series_repo
-            .find_by_id(series_id)
-            .await
-            .unwrap()
-            .unwrap();
-        let target = start + Duration::weeks(10);
-        let added = svc
-            .recurring_event_service
-            .extend_horizon(&series, target)
-            .await
-            .unwrap();
-        assert_eq!(
-            added, 5,
-            "stream positions 6..=10 should all be materialized (none skipped)"
-        );
-
-        // (a) The occurrence at stream position N+1 = 6 IS created — not
-        // skipped by colliding with occurrence 5's Cancelled exception —
-        // and it lands at the correct stream date (anchor + 5 weeks).
-        let next = svc
-            .event_repo
-            .find_by_series_and_index(series_id, 6)
-            .await
-            .unwrap()
-            .expect("occurrence at stream position 6 must be created, not skipped");
-        assert_eq!(
-            next.start_time,
-            start + Duration::weeks(5),
-            "occurrence_index 6 must sit at stream position 6 (anchor + 5 weeks)"
-        );
-
-        // (b) Every newly created row carries a stream-aligned index:
-        // index k ⇔ start_time = anchor + (k-1) weeks. Under the old
-        // MAX-based numbering, index 6 would point at week 6 (the skip
-        // shifts each later row's date), so this catches the regression.
-        for idx in 6..=10 {
-            let occ = svc
-                .event_repo
-                .find_by_series_and_index(series_id, idx)
-                .await
-                .unwrap()
-                .unwrap_or_else(|| panic!("occurrence_index {idx} should exist"));
-            assert_eq!(
-                occ.start_time,
-                start + Duration::weeks((idx - 1) as i64),
-                "occurrence_index {idx} must align with its stream position",
-            );
-        }
-
-        // Full index set: survivors 1..=4 (5 was cancelled) plus the new
-        // 6..=10 — exactly the stream positions, no downward shift.
-        let mut indices: Vec<i32> = sqlx::query_scalar(
-            "SELECT occurrence_index FROM events WHERE series_id = ? ORDER BY occurrence_index ASC",
-        )
-        .bind(series_id.to_string())
-        .fetch_all(&pool)
-        .await
-        .unwrap();
-        indices.sort();
-        assert_eq!(indices, vec![1, 2, 3, 4, 6, 7, 8, 9, 10]);
-
-        // The cancelled boundary index stays cancelled.
-        assert!(svc
-            .event_series_repo
-            .find_exception(series_id, 5)
-            .await
-            .unwrap()
-            .is_some());
-    }
-
-    /// Regression for the reviewer's anchor-shift concern: cancelling the
-    /// FIRST occurrence hard-deletes its row, so `MIN(start_time)` (the
-    /// re-derived anchor) jumps to occurrence #2. The base index must be
-    /// offset by the earliest survivor's true `occurrence_index`,
-    /// otherwise a naive count from the shifted anchor undercounts by one
-    /// and the first new occurrence collides with the surviving
-    /// occurrence #5's index — corrupting the stream numbering.
-    #[tokio::test]
-    async fn extend_horizon_after_cancelling_first_occurrence_keeps_indices_aligned() {
-        let pool = fresh_pool().await;
-        let svc = make_service(pool.clone());
-        let actor = make_actor(&pool).await;
-
-        // Materialize EXACTLY occurrences 1..=5 (N = 5 is the boundary).
-        let start = next_tuesday_anchor();
-        let until = start + Duration::weeks(4) + Duration::days(1);
-        let anchor = svc
-            .create(actor, recurring_input(start, until))
-            .await
-            .unwrap();
-        let series_id = anchor.series_id.unwrap();
-        let initial: (i64,) = sqlx::query_as("SELECT COUNT(*) FROM events WHERE series_id = ?")
-            .bind(series_id.to_string())
-            .fetch_one(&pool)
-            .await
-            .unwrap();
-        assert_eq!(initial.0, 5);
-
-        // Open the series back up so extend_horizon can roll past the cap.
-        sqlx::query("UPDATE event_series SET until_date = NULL WHERE id = ?")
-            .bind(series_id.to_string())
-            .execute(&pool)
-            .await
-            .unwrap();
-
-        // Cancel occurrence #1: deletes its events row (so MIN(start_time)
-        // now resolves to occurrence #2's start), writes a Cancelled
-        // exception at index 1, leaves materialized_through unchanged.
-        svc.cancel_event_occurrence(actor, series_id, 1, Some("first".into()))
-            .await
-            .unwrap();
-        assert!(
-            svc.event_repo
-                .find_by_series_and_index(series_id, 1)
-                .await
-                .unwrap()
-                .is_none(),
-            "occurrence 1 should be deleted after cancel"
-        );
-
-        // Roll the horizon forward to cover stream positions 6..=10.
-        let series = svc
-            .event_series_repo
-            .find_by_id(series_id)
-            .await
-            .unwrap()
-            .unwrap();
-        let target = start + Duration::weeks(10);
-        let added = svc
-            .recurring_event_service
-            .extend_horizon(&series, target)
-            .await
-            .unwrap();
-        assert_eq!(
-            added, 5,
-            "stream positions 6..=10 should all be materialized"
-        );
-
-        // The first new occurrence lands at stream position 6 (anchor + 5
-        // weeks) and carries occurrence_index 6 — NOT 5. Under the buggy
-        // MIN-anchor count it would be assigned index 5, colliding with
-        // the surviving occurrence #5.
-        let next = svc
-            .event_repo
-            .find_by_series_and_index(series_id, 6)
-            .await
-            .unwrap()
-            .expect("occurrence at stream position 6 must be created");
-        assert_eq!(
-            next.start_time,
-            start + Duration::weeks(5),
-            "occurrence_index 6 must sit at stream position 6 (anchor + 5 weeks)"
-        );
-
-        // Every row aligns with its stream position: index k ⇔ start_time
-        // = anchor + (k-1) weeks, for the survivors (2..=5) and the new
-        // batch (6..=10) alike.
-        for idx in 2..=10 {
-            let occ = svc
-                .event_repo
-                .find_by_series_and_index(series_id, idx)
-                .await
-                .unwrap()
-                .unwrap_or_else(|| panic!("occurrence_index {idx} should exist"));
-            assert_eq!(
-                occ.start_time,
-                start + Duration::weeks((idx - 1) as i64),
-                "occurrence_index {idx} must align with its stream position",
-            );
-        }
-
-        // Full index set: survivors 2..=5 (1 was cancelled) plus the new
-        // 6..=10 — exactly the stream positions, no collision, no shift.
-        let mut indices: Vec<i32> = sqlx::query_scalar(
-            "SELECT occurrence_index FROM events WHERE series_id = ? ORDER BY occurrence_index ASC",
-        )
-        .bind(series_id.to_string())
-        .fetch_all(&pool)
-        .await
-        .unwrap();
-        indices.sort();
-        assert_eq!(indices, vec![2, 3, 4, 5, 6, 7, 8, 9, 10]);
-
-        // The cancelled first index stays cancelled.
-        assert!(svc
-            .event_series_repo
-            .find_exception(series_id, 1)
-            .await
-            .unwrap()
-            .is_some());
-    }
-
-    /// Regression guard for the new stream-position base-index: with NO
-    /// cancellations, extend_horizon must still continue the prior batch
-    /// contiguously (no off-by-one, no gap, no duplicate index).
-    #[tokio::test]
-    async fn extend_horizon_with_no_cancellations_assigns_contiguous_indices() {
-        let pool = fresh_pool().await;
-        let svc = make_service(pool.clone());
-        let actor = make_actor(&pool).await;
-
-        // Materialize occurrences 1..=5, then extend with NO cancellations.
-        let start = next_tuesday_anchor();
-        let until = start + Duration::weeks(4) + Duration::days(1);
-        let anchor = svc
-            .create(actor, recurring_input(start, until))
-            .await
-            .unwrap();
-        let series_id = anchor.series_id.unwrap();
-        let initial: (i64,) = sqlx::query_as("SELECT COUNT(*) FROM events WHERE series_id = ?")
-            .bind(series_id.to_string())
-            .fetch_one(&pool)
-            .await
-            .unwrap();
-        assert_eq!(initial.0, 5);
-
-        sqlx::query("UPDATE event_series SET until_date = NULL WHERE id = ?")
-            .bind(series_id.to_string())
-            .execute(&pool)
-            .await
-            .unwrap();
-
-        let series = svc
-            .event_series_repo
-            .find_by_id(series_id)
-            .await
-            .unwrap()
-            .unwrap();
-        let target = start + Duration::weeks(10);
-        let added = svc
-            .recurring_event_service
-            .extend_horizon(&series, target)
-            .await
-            .unwrap();
-        assert_eq!(added, 5, "occurrences 6..=10 should be materialized");
-
-        // The new batch continues contiguously from the prior one: the
-        // full set is 1..=10 with no gap, no duplicate, and no off-by-one
-        // from the stream-position base-index computation.
-        let mut indices: Vec<i32> = sqlx::query_scalar(
-            "SELECT occurrence_index FROM events WHERE series_id = ? ORDER BY occurrence_index ASC",
-        )
-        .bind(series_id.to_string())
-        .fetch_all(&pool)
-        .await
-        .unwrap();
-        indices.sort();
-        assert_eq!(indices, (1..=10).collect::<Vec<i32>>());
-
-        // And each row aligns with its stream position.
-        for idx in 1..=10 {
-            let occ = svc
-                .event_repo
-                .find_by_series_and_index(series_id, idx)
-                .await
-                .unwrap()
-                .unwrap_or_else(|| panic!("occurrence_index {idx} should exist"));
-            assert_eq!(occ.start_time, start + Duration::weeks((idx - 1) as i64));
-        }
-    }
-
-    #[tokio::test]
-    async fn override_event_occurrence_updates_row_and_writes_exception() {
-        let pool = fresh_pool().await;
-        let svc = make_service(pool.clone());
-        let actor = make_actor(&pool).await;
-        let (series_id, _anchor) = make_series_with_anchor(&svc, actor).await;
-
-        let ov = OccurrenceOverride {
-            location: Some("Conference Room B".into()),
-            ..Default::default()
-        };
-        let updated = svc
-            .override_event_occurrence(actor, series_id, 5, ov, None)
-            .await
-            .unwrap();
-        assert_eq!(updated.location.as_deref(), Some("Conference Room B"));
-
-        // The events row reflects the override.
-        let fetched = svc
-            .event_repo
-            .find_by_series_and_index(series_id, 5)
-            .await
-            .unwrap()
-            .unwrap();
-        assert_eq!(fetched.location.as_deref(), Some("Conference Room B"));
-
-        // Exception row exists with the JSON payload.
-        let ex = svc
-            .event_series_repo
-            .find_exception(series_id, 5)
-            .await
-            .unwrap()
-            .unwrap();
-        assert_eq!(ex.kind, OccurrenceExceptionKind::Overridden);
-        let payload = ex.override_payload.expect("override payload");
-        assert!(payload.contains("Conference Room B"));
-
-        // Audit row.
-        assert!(
-            audit_count(&pool, "override_event_occurrence", &updated.id.to_string()).await >= 1,
-        );
-    }
-
-    #[tokio::test]
-    async fn overridden_occurrence_survives_series_edit() {
-        let pool = fresh_pool().await;
-        let svc = make_service(pool.clone());
-        let actor = make_actor(&pool).await;
-        let (series_id, anchor) = make_series_with_anchor(&svc, actor).await;
-
-        // Override occurrence 3's location.
-        let ov = OccurrenceOverride {
-            location: Some("Room B".into()),
-            ..Default::default()
-        };
-        svc.override_event_occurrence(actor, series_id, 3, ov, None)
-            .await
-            .unwrap();
-
-        // Update the series template (change-future) starting from
-        // occurrence 1 — this propagates a new location to all rows in
-        // the series.
-        let mut update = update_input_from(&anchor);
-        update.location = Some("Old Room A".into());
-        update.title = "Edited Title".into();
-        svc.update_series_from(actor, series_id, anchor.start_time, update)
-            .await
-            .unwrap();
-
-        // Other occurrences picked up the series-level location...
-        let other = svc
-            .event_repo
-            .find_by_series_and_index(series_id, 2)
-            .await
-            .unwrap()
-            .unwrap();
-        assert_eq!(other.location.as_deref(), Some("Old Room A"));
-        assert_eq!(other.title, "Edited Title");
-
-        // ...but the overridden one retains its per-occurrence value.
-        // update_series_from currently rewrites every row including
-        // overridden ones; the override must be re-applied to model
-        // "override wins." Re-apply via the materializer's exception
-        // path by calling extend_horizon (which re-applies overrides
-        // on re-creation). For now we assert the simpler invariant:
-        // the exception row is still recorded so future materializations
-        // re-apply the override.
-        let ex = svc
-            .event_series_repo
-            .find_exception(series_id, 3)
-            .await
-            .unwrap();
-        assert!(
-            ex.is_some(),
-            "override exception should survive series edit"
-        );
-        assert_eq!(ex.unwrap().kind, OccurrenceExceptionKind::Overridden);
-    }
-
-    #[tokio::test]
-    async fn restore_cancelled_recreates_row() {
-        let pool = fresh_pool().await;
-        let svc = make_service(pool.clone());
-        let actor = make_actor(&pool).await;
-        let (series_id, _) = make_series_with_anchor(&svc, actor).await;
-
-        // Capture the original start_time of occurrence 6 before we
-        // cancel it.
-        let original = svc
-            .event_repo
-            .find_by_series_and_index(series_id, 6)
-            .await
-            .unwrap()
-            .unwrap();
-        let original_start = original.start_time;
-
-        svc.cancel_event_occurrence(actor, series_id, 6, None)
-            .await
-            .unwrap();
-        assert!(svc
-            .event_repo
-            .find_by_series_and_index(series_id, 6)
-            .await
-            .unwrap()
-            .is_none());
-
-        let restored = svc
-            .restore_event_occurrence(actor, series_id, 6)
-            .await
-            .unwrap();
-        let restored = restored.expect("cancel-restore returns Some(event)");
-        assert_eq!(restored.occurrence_index, Some(6));
-        assert_eq!(restored.start_time, original_start);
-        assert!(svc
-            .event_series_repo
-            .find_exception(series_id, 6)
-            .await
-            .unwrap()
-            .is_none());
-    }
-
-    #[tokio::test]
-    async fn restore_overridden_resets_to_template() {
-        let pool = fresh_pool().await;
-        let svc = make_service(pool.clone());
-        let actor = make_actor(&pool).await;
-        let (series_id, _) = make_series_with_anchor(&svc, actor).await;
-
-        let original = svc
-            .event_repo
-            .find_by_series_and_index(series_id, 4)
-            .await
-            .unwrap()
-            .unwrap();
-        let original_title = original.title.clone();
-
-        let ov = OccurrenceOverride {
-            title: Some("Special One-off".into()),
-            ..Default::default()
-        };
-        svc.override_event_occurrence(actor, series_id, 4, ov, None)
-            .await
-            .unwrap();
-        let after_override = svc
-            .event_repo
-            .find_by_series_and_index(series_id, 4)
-            .await
-            .unwrap()
-            .unwrap();
-        assert_eq!(after_override.title, "Special One-off");
-
-        let result = svc
-            .restore_event_occurrence(actor, series_id, 4)
-            .await
-            .unwrap();
-        assert!(result.is_none(), "override-restore returns None");
-
-        let restored = svc
-            .event_repo
-            .find_by_series_and_index(series_id, 4)
-            .await
-            .unwrap()
-            .unwrap();
-        assert_eq!(restored.title, original_title);
-        assert!(svc
-            .event_series_repo
-            .find_exception(series_id, 4)
-            .await
-            .unwrap()
-            .is_none());
-    }
-
-    #[tokio::test]
-    async fn cancel_then_cancel_is_idempotent() {
-        let pool = fresh_pool().await;
-        let svc = make_service(pool.clone());
-        let actor = make_actor(&pool).await;
-        let (series_id, _) = make_series_with_anchor(&svc, actor).await;
-
-        svc.cancel_event_occurrence(actor, series_id, 2, Some("first".into()))
-            .await
-            .unwrap();
-        // Second call should succeed without error.
-        svc.cancel_event_occurrence(actor, series_id, 2, Some("second".into()))
-            .await
-            .unwrap();
-
-        let ex = svc
-            .event_series_repo
-            .find_exception(series_id, 2)
-            .await
-            .unwrap()
-            .unwrap();
-        assert_eq!(ex.kind, OccurrenceExceptionKind::Cancelled);
-        // The second insert (via UPSERT) overwrites the reason.
-        assert_eq!(ex.audit_reason.as_deref(), Some("second"));
-
-        // No events row.
-        assert!(svc
-            .event_repo
-            .find_by_series_and_index(series_id, 2)
-            .await
-            .unwrap()
-            .is_none());
-    }
-
-    #[tokio::test]
-    async fn audit_rows_emitted_for_each_action() {
-        let pool = fresh_pool().await;
-        let svc = make_service(pool.clone());
-        let actor = make_actor(&pool).await;
-        let (series_id, _) = make_series_with_anchor(&svc, actor).await;
-
-        let before = svc
-            .event_repo
-            .find_by_series_and_index(series_id, 7)
-            .await
-            .unwrap()
-            .unwrap();
-
-        svc.cancel_event_occurrence(actor, series_id, 7, None)
-            .await
-            .unwrap();
-        assert!(audit_count(&pool, "cancel_event_occurrence", &before.id.to_string()).await >= 1,);
-
-        svc.restore_event_occurrence(actor, series_id, 7)
-            .await
-            .unwrap();
-        assert!(
-            audit_count(
-                &pool,
-                "restore_event_occurrence",
-                &format!("{}#7", series_id)
-            )
-            .await
-                >= 1,
-        );
-
-        let ov = OccurrenceOverride {
-            location: Some("Room Z".into()),
-            ..Default::default()
-        };
-        let updated = svc
-            .override_event_occurrence(actor, series_id, 8, ov, None)
-            .await
-            .unwrap();
-        assert!(
-            audit_count(&pool, "override_event_occurrence", &updated.id.to_string()).await >= 1,
-        );
-    }
-
-    /// Bonus: round-trip the materializer's exception-aware path. Use
-    /// an open-ended series (no until_date) so extend_horizon can roll
-    /// forward past the initial materialization window, and seed an
-    /// override exception at a future-index slot. The materializer
-    /// should apply the override when it eventually creates that row.
-    #[tokio::test]
-    async fn materializer_re_applies_override_on_extend() {
-        let pool = fresh_pool().await;
-        let svc = make_service(pool.clone());
-        let actor = make_actor(&pool).await;
-
-        // Open-ended series.
-        let start = next_tuesday_anchor();
-        let mut input = single_input(start, EventVisibility::MembersOnly);
-        input.recurrence = Some(Recurrence::WeeklyByDay {
-            interval: 1,
-            weekdays: vec![WeekdayCode::Tue],
-        });
-        input.recurrence_until = None;
-        let anchor = svc.create(actor, input).await.unwrap();
-        let series_id = anchor.series_id.unwrap();
-
-        // Find the current max materialized index. Pick one just
-        // beyond it as our target exception slot.
-        let current_max = svc
-            .event_repo
-            .max_occurrence_index_for_series(series_id)
-            .await
-            .unwrap()
-            .unwrap_or(0);
-        let target_index = current_max + 5;
-
-        let payload = serde_json::to_string(&OccurrenceOverride {
-            location: Some("Far Future Room".into()),
-            ..Default::default()
-        })
-        .unwrap();
-        svc.event_series_repo
-            .insert_exception(OccurrenceException {
-                series_id,
-                occurrence_index: target_index,
-                kind: OccurrenceExceptionKind::Overridden,
-                override_payload: Some(payload),
-                created_at: Utc::now(),
-                created_by: actor,
-                audit_reason: None,
-            })
-            .await
-            .unwrap();
-
-        // Extend the horizon enough to cover the target slot — for a
-        // weekly series, +5 weeks is enough for 5 more occurrences.
-        let series = svc
-            .event_series_repo
-            .find_by_id(series_id)
-            .await
-            .unwrap()
-            .unwrap();
-        let far = series.materialized_through + chrono::Duration::weeks(6);
-        svc.recurring_event_service
-            .extend_horizon(&series, far)
-            .await
-            .unwrap();
-
-        let row = svc
-            .event_repo
-            .find_by_series_and_index(series_id, target_index)
-            .await
-            .unwrap();
-        let row = row.expect("target occurrence should have been materialized");
-        assert_eq!(row.location.as_deref(), Some("Far Future Room"));
-    }
-
-    #[tokio::test]
-    async fn delete_series_cascades_and_audits() {
-        let pool = fresh_pool().await;
-        let svc = make_service(pool.clone());
-        let actor = make_actor(&pool).await;
-
-        let start = next_tuesday_anchor();
-        let until = weeks_after(start, 8);
-        let mut input = single_input(start, EventVisibility::MembersOnly);
-        input.recurrence = Some(Recurrence::WeeklyByDay {
-            interval: 1,
-            weekdays: vec![WeekdayCode::Tue],
-        });
-        input.recurrence_until = Some(until);
-
-        let anchor = svc.create(actor, input).await.unwrap();
-        let series_id = anchor.series_id.unwrap();
-
-        svc.delete_series(actor, series_id).await.unwrap();
-
-        // Series row gone.
-        assert!(svc
-            .event_series_repo
-            .find_by_id(series_id)
-            .await
-            .unwrap()
-            .is_none());
-        // All occurrences cascade-deleted.
-        let count: (i64,) = sqlx::query_as("SELECT COUNT(*) FROM events WHERE series_id = ?")
-            .bind(series_id.to_string())
-            .fetch_one(&pool)
-            .await
-            .unwrap();
-        assert_eq!(count.0, 0);
-
-        assert_eq!(
-            audit_count(&pool, "delete_event_series", &series_id.to_string()).await,
-            1,
-        );
-    }
-
-    // -----------------------------------------------------------------
-    // Error-path tests — assert typed AppError variants on operator-
-    // reachable failure inputs (stale UUIDs, hand-crafted zero-index
-    // URLs) and the no-op + audit branch of restore.
-
-    #[tokio::test]
-    async fn update_one_errors_when_event_id_not_found() {
-        let pool = fresh_pool().await;
-        let svc = make_service(pool.clone());
-        let actor = make_actor(&pool).await;
-
-        let missing_id = Uuid::new_v4();
-        let input = UpdateEventInput {
-            title: "Whatever".to_string(),
-            description: "Body".to_string(),
-            event_type: EventType::Meeting,
-            event_type_id: None,
-            visibility: EventVisibility::MembersOnly,
-            start_time: next_saturday_anchor(),
-            end_time: None,
-            location: None,
-            max_attendees: None,
-            rsvp_required: false,
-            image_url: None,
-        };
-
-        let err = svc.update_one(actor, missing_id, input).await.unwrap_err();
-        match err {
-            AppError::NotFound(msg) => assert!(
-                msg.contains("Event not found"),
-                "expected 'Event not found' in message, got: {msg}",
-            ),
-            other => panic!("expected NotFound, got {other:?}"),
-        }
-
-        // Short-circuit BEFORE the audit row — no phantom update_event entry.
-        assert_eq!(
-            audit_count(&pool, "update_event", &missing_id.to_string()).await,
-            0,
-        );
-    }
-
-    #[tokio::test]
-    async fn cancel_event_occurrence_errors_when_series_not_found() {
-        let pool = fresh_pool().await;
-        let svc = make_service(pool.clone());
-        let actor = make_actor(&pool).await;
-
-        let missing_series = Uuid::new_v4();
-        let err = svc
-            .cancel_event_occurrence(actor, missing_series, 1, None)
-            .await
-            .unwrap_err();
-        match err {
-            AppError::NotFound(msg) => {
-                assert!(msg.contains("series"), "expected 'series' in msg: {msg}");
-                assert!(
-                    msg.contains("not found"),
-                    "expected 'not found' in msg: {msg}",
-                );
-            }
-            other => panic!("expected NotFound, got {other:?}"),
-        }
-    }
-
-    #[tokio::test]
-    async fn cancel_event_occurrence_errors_on_zero_index() {
-        let pool = fresh_pool().await;
-        let svc = make_service(pool.clone());
-        let actor = make_actor(&pool).await;
-        let (series_id, _) = make_series_with_anchor(&svc, actor).await;
-
-        let err = svc
-            .cancel_event_occurrence(actor, series_id, 0, None)
-            .await
-            .unwrap_err();
-        match err {
-            AppError::BadRequest(msg) => assert!(
-                msg.contains("occurrence_index must be >= 1"),
-                "expected zero-index msg, got: {msg}",
-            ),
-            other => panic!("expected BadRequest, got {other:?}"),
-        }
-    }
-
-    #[tokio::test]
-    async fn override_event_occurrence_errors_when_series_not_found() {
-        let pool = fresh_pool().await;
-        let svc = make_service(pool.clone());
-        let actor = make_actor(&pool).await;
-
-        let missing_series = Uuid::new_v4();
-        let ov = OccurrenceOverride {
-            location: Some("Anywhere".into()),
-            ..Default::default()
-        };
-        let err = svc
-            .override_event_occurrence(actor, missing_series, 1, ov, None)
-            .await
-            .unwrap_err();
-        match err {
-            AppError::NotFound(msg) => {
-                assert!(msg.contains("series"), "expected 'series' in msg: {msg}");
-                assert!(
-                    msg.contains("not found"),
-                    "expected 'not found' in msg: {msg}",
-                );
-            }
-            other => panic!("expected NotFound, got {other:?}"),
-        }
-    }
-
-    #[tokio::test]
-    async fn override_event_occurrence_errors_on_zero_index() {
-        let pool = fresh_pool().await;
-        let svc = make_service(pool.clone());
-        let actor = make_actor(&pool).await;
-        let (series_id, _) = make_series_with_anchor(&svc, actor).await;
-
-        let ov = OccurrenceOverride {
-            location: Some("Anywhere".into()),
-            ..Default::default()
-        };
-        let err = svc
-            .override_event_occurrence(actor, series_id, 0, ov, None)
-            .await
-            .unwrap_err();
-        match err {
-            AppError::BadRequest(msg) => assert!(
-                msg.contains("occurrence_index must be >= 1"),
-                "expected zero-index msg, got: {msg}",
-            ),
-            other => panic!("expected BadRequest, got {other:?}"),
-        }
-    }
-
-    #[tokio::test]
-    async fn restore_event_occurrence_errors_when_series_not_found() {
-        let pool = fresh_pool().await;
-        let svc = make_service(pool.clone());
-        let actor = make_actor(&pool).await;
-
-        // Hits the inline find_by_id-then-ok_or_else chain in
-        // restore_event_occurrence, not require_series_exists.
-        let missing_series = Uuid::new_v4();
-        let err = svc
-            .restore_event_occurrence(actor, missing_series, 1)
-            .await
-            .unwrap_err();
-        match err {
-            AppError::NotFound(msg) => {
-                assert!(msg.contains("series"), "expected 'series' in msg: {msg}");
-            }
-            other => panic!("expected NotFound, got {other:?}"),
-        }
-    }
-
-    #[tokio::test]
-    async fn restore_event_occurrence_noop_when_no_exception_emits_audit() {
-        let pool = fresh_pool().await;
-        let svc = make_service(pool.clone());
-        let actor = make_actor(&pool).await;
-        let (series_id, _) = make_series_with_anchor(&svc, actor).await;
-
-        // Occurrence 2 exists (materializer creates several) and has no
-        // exception — restore must be a no-op but still audit.
-        let result = svc
-            .restore_event_occurrence(actor, series_id, 2)
-            .await
-            .unwrap();
-        assert!(result.is_none(), "no-op restore returns None");
-
-        assert_eq!(
-            audit_count(
-                &pool,
-                "restore_event_occurrence",
-                &format!("{}#2", series_id),
-            )
-            .await,
-            1,
-        );
-    }
-}
+#[path = "event_admin_service_tests.rs"]
+mod tests;
